获取幸运值api返回为空和计算总算力bug修改

+++ models/models.go
@@ -137,7 +137,7 @@ type AllMinersMined struct {
 	Epoch      int64  `orm:"index"`
 	Reward     float64
 	Power      float64 `orm:"index"`
-	TotalPower float64
+	TotalPower int64
 	Time       time.Time
 }
